+++ packages/dart_pdf_editor_flutter_gpu/lib/src/backend_stub.dart
@@ -1,6 +1,7 @@
 import 'package:dart_pdf_editor/dart_pdf_editor.dart';
 
 import 'backend_stats.dart';
+import 'text_outliner.dart';
 
 /// Web/unsupported compile-time stub for [FlutterGpuTileRasterBackend].
 class FlutterGpuTileRasterBackend extends PdfTileRasterBackend {
@@ -10,6 +11,8 @@ class FlutterGpuTileRasterBackend extends PdfTileRasterBackend {
     this.maxTextureBytes = 256 << 20,
     this.maxGeometryBytes = 256 << 20,
     this.enableProactiveWarmUp,
+    this.textOutliner,
+    this.systemTextOutlines = false,
     FlutterGpuTileBackendStats? stats,
   }) : stats = stats ?? FlutterGpuTileBackendStats();
 
@@ -18,6 +21,8 @@ class FlutterGpuTileRasterBackend extends PdfTileRasterBackend {
   final int maxTextureBytes;
   final int maxGeometryBytes;
   final bool? enableProactiveWarmUp;
+  final FlutterGpuTextOutliner? textOutliner;
+  final bool systemTextOutlines;
   final FlutterGpuTileBackendStats stats;
   String? _lastSessionRejection;