+++ packages/dart_pdf_editor_flutter_gpu/lib/src/system_text_outliner_stub.dart
@@ -1,0 +1,13 @@
+import 'package:pdf_graphics/pdf_graphics.dart';
+
+import 'text_outliner.dart';
+
+/// Web stub for the native system-font outline adapter.
+class FlutterGpuSystemTextOutliner implements FlutterGpuTextOutliner {
+  FlutterGpuSystemTextOutliner._();
+
+  static FlutterGpuSystemTextOutliner? tryCreate() => null;
+
+  @override
+  PdfTextRun? outline(PdfTextRun run) => null;
+}